WIP - deleted criteria, see description
We have removed criteria from this library in lieu of laravel's supported query scopes. Query scopes are a better, more officially known version of this pattern than anything we had implemented here previously, so it makes sense to standardize and use them instead of having to maintain this separate version.